Ajax Image Upload
Ajax Image upload done. In users folder the images is changing but can't
get the DOM elements and make changes in it. Vivek please look at the
uploadimg.php code for it please.I don't know if i have DOM extension,
maybe that is the problem. You please look at it.